requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
Faker==21.0.0
idna==3.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
